refactor: add check suffix for check files
